Removed bootstrap.js. Fixed regex error in qpanda.js.
Bootstrap was only needed for when the user dismisses an error notification. That functionality was replaced with 3 lines of code in qpanda.js. Before I was using an <a> element to close the notification and had href="#" which did nothing. But that was dumb. I could have used a button, which I changed it too. I read a stackoverflow answer that used an <a> tag and used it in mine. Also the regex pattern matching wasn't working during the register form check, fixed now.