Logging in works. Renamed login view. Added new bandwidth consideration.
Have a basic login that works. Renamed the login view to user_login for readability because we call a django function also called login from within that view. Added consideration about removing an unnecessary input field for login that is sent in a POST request even when it is empty.