Redirects to appropriate page on login/logout. (Doesn't work on register yet)
Using get parameters in the form action we can now redirect the user to an appropriate page on login/logout. For example both askquestion and askedquestion use the registerlogin template. However they both call the login view, we use the get parameters to redirect to the view that called it. If you login viewing a question you will return to that question logged in.