Added a register and login page for when qpanda.co/register/ and qpanda.co/login/ are called. Added a check for the URL rerouting.
Added a consideration about whether to support javascript-less clients. Added template that supports javascript-less registration (which I need regardless of which direction I take after I decide on Consideration 3b). Also added a more comprehensive check for which URLs to redirect to in nextURL. I don't redirect to views which are mainly used for form submits. Changed the confirmpasswordfield (in registerloginform) to include a class so that the register button (in hardregister) still works using the same jquery click function.