The AJAX get request is now dynamic.
The get request will call for questions beyond the one it already has displayed instead of using the static 10 we had before. The number of the last answer is stored inside a data-fromanswers attr and is used to calculate which answers to ask for in the get request. Also if you click load more answers several times and reload the page, it will load the answers you were last up to (by using get parameters) instead of having to manually click the button.