Display a tooltip in the confirmpasswordfield when form is invalid. Register button takes up full width.
The input field will display an error message in a tooltip that will explain why the red x is appearing when the user is entering their details. If the user clicks register we just display, we hide the login button. I still need to have a login button appear in case the user accidentally clicks register but wished to login.