Doesn't display a button to load more answers when there are no more answers to be displayed.
If a user is viewing the last of the answers we don't display the getmoreanswers button which sends unneeded AJAX requests to the server. This only works with requests directly to the server e.g. qpanda.co/1234567/?from=10. We still need to hide the button after an AJAX request for the last answer has been made.